Create README.md
Magento Marketplace Responsive Template is an add-on that will support all the screens you are view a store in marketplace website on iPhone, iPad, Android-based mobile devices or small or large screen devices, It will work well in all screens. After installing this theme, the list of market sellers and their products will be displayed on the homepage.